requirements.txt
discord.py==1.7.3
requests==2.22.0
Django==3.2.6
beautifulsoup4==4.9.3
discord==1.7.3
djangorestframework==3.12.4
google_api_python_client==2.18.0
python-dotenv==0.19.0
gunicorn==20.1.0